--- web/src/engine/websites/TecnoScan.ts
@@ -1,19 +1,0 @@
-import { Tags } from '../Tags';
-import icon from './TecnoScan.webp';
-import { DecoratableMangaScraper } from '../providers/MangaPlugin';
-import * as Madara from './decorators/WordPressMadara';
-import * as Common from './decorators/Common';
-
-@Madara.MangaCSS(/^{origin}\/manga\/[^/]+\/$/, 'ol.breadcrumb li:last-of-type a')
-@Madara.MangasMultiPageAJAX()
-@Madara.ChaptersSinglePageAJAXv2()
-@Madara.PagesSinglePageCSS()
-@Common.ImageAjax()
-export default class extends DecoratableMangaScraper {
-    public constructor() {
-        super('tecnoscan', 'Tecno Scan', 'https://visortecno.com', Tags.Media.Manhua, Tags.Media.Manhwa, Tags.Language.Spanish, Tags.Source.Aggregator);
-    }
-    public override get Icon() {
-        return icon;
-    }
-}
