+++ web/src/engine/websites/_index.ts
@@ -656,7 +656,6 @@ export { default as TarotScans } from './TarotScans';
 export { default as TatakaeScan } from './TatakaeScan';
 export { default as TCBScans } from './TCBScans';
 export { default as Team1x1 } from './Team1x1';
-export { default as TecnoScan } from './TecnoScan';
 export { default as Tempestfansub } from './Tempestfansub';
 export { default as TempestScans } from './TempestScans';
 export { default as TenshiID } from './TenshiID';
@@ -679,6 +678,7 @@ export { default as TortugaCeviri } from './TortugaCeviri';
 export { default as Toti } from './Toti';
 export { default as TraduccionesAmistosas } from './TraduccionesAmistosas';
 export { default as TraduccionesMoonlight } from './TraduccionesMoonlight';
+export { default as TresDaos } from './TresDaos';
 export { default as TritiniaScans } from './TritiniaScans';
 export { default as TruyenQQ } from './TruyenQQ';
 export { default as TruyenTranhDammy } from './TruyenTranhDammy';
